remove unused views restructure files
