+++ src/js/components/app-container.jsx
@@ -3,8 +3,8 @@ import {BrowserRouter as Router, Route, Switch} from 'react-router-dom';
 import {Provider} from 'react-redux';
 import PropTypes from 'prop-types';
 import WalletView from './wallet-view';
-import ImportWalletView from './import-wallet-view';
-import UnlockWalletView from './unlock-wallet-view';
+import ImportWalletView from './auth/import-wallet-view';
+import UnlockWalletView from './auth/unlock-wallet-view';
 import UnlockedWalletRequiredRoute from './utils/unlocked-wallet-required-route';
 import TransactionHistoryView from './transaction-history-view';
 import UnspentTransactionOutputView from './unspent-transaction-output-view';
@@ -14,7 +14,7 @@ import PeerInfoView from './peer-info-view';
 import AdvertisementFormView from './advertisement/advertisement-form-view';
 import AdvertisementListView from './advertisement/advertisement-list-view';
 import ActionView from './action-view';
-import NewWalletView from './new-wallet-view';
+import NewWalletView from './auth/new-wallet-view';
 import ManageWalletView from './manage-wallet-view';
 import StatsView from './stats-view';
 import BacklogView from './backlog-view';
