+++ src/js/components/auth/import-wallet-view.jsx
@@ -2,25 +2,21 @@ import React, {Component} from 'react';
 import {connect} from 'react-redux';
 import {Redirect} from 'react-router-dom';
 import {Button, Col, Container, Row} from 'react-bootstrap';
-import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
-import API from '../api/index';
-import WalletCreatePasswordView from './utils/wallet-create-password-view';
-import MnemonicView from './utils/mnemonic-view';
-import MnemonicConfirmView from './utils/mnemonic-confirm-view';
-import NewWalletStepProgressView from './utils/new-wallet-step-progress-view';
-import WalletCreateInfoView from './utils/wallet-create-info-view';
-import {unlockWallet} from '../redux/actions';
-import Translation from '../common/translation';
+import API from '../../api';
+import WalletCreatePasswordView from './wallet-create-password-view';
+import MnemonicConfirmView from './mnemonic-confirm-view';
+import WalletCreateInfoView from './wallet-create-info-view';
+import {unlockWallet} from '../../redux/actions';
+import Translation from '../../common/translation';
 
 const STATUS = {
-    NEW_WALLET_PASSWORD        : 1,
-    NEW_WALLET_MNEMONIC        : 2,
-    NEW_WALLET_CONFIRM_MNEMONIC: 3,
-    NEW_WALLET_CREATED         : 4
+    NEW_WALLET_MNEMONIC: 1,
+    NEW_WALLET_PASSWORD: 2,
+    NEW_WALLET_CREATED : 3
 };
 
 
-class NewWalletView extends Component {
+class ImportWalletView extends Component {
     constructor(props) {
         super(props);
         this.state = {
@@ -30,15 +26,11 @@ class NewWalletView extends Component {
             password_confirm     : undefined,
             password_valid       : false,
             wallet_info          : undefined,
-            status               : STATUS.NEW_WALLET_PASSWORD
+            status               : STATUS.NEW_WALLET_MNEMONIC
         };
     }
 
     componentDidMount() {
-        API.getRandomMnemonic()
-           .then(data => {
-               this.setState({mnemonic: data.mnemonic.split(' ')});
-           });
     }
 
     verifyPassword() {
@@ -61,34 +53,17 @@ class NewWalletView extends Component {
            });
     }
 
-    getStepName() {
-        const result_step = [
-            {
-                label: Translation.getPhrase('022a0c60c')
-            },
-            {
-                label: Translation.getPhrase('5c14bcece')
-            },
-            {
-                label: Translation.getPhrase('a22a090ac')
-            },
-            {
-                label: Translation.getPhrase('b192f3b66')
-            }
-        ];
-
-        return result_step[this.state.status - 1].label;
-    }
-
     createWalletNextStep() {
         switch (this.state.status) {
-            case STATUS.NEW_WALLET_PASSWORD:
-                this.setState({status: STATUS.NEW_WALLET_MNEMONIC});
-                break;
             case STATUS.NEW_WALLET_MNEMONIC:
-                this.setState({status: STATUS.NEW_WALLET_CONFIRM_MNEMONIC});
+                this.setState({
+                    status          : STATUS.NEW_WALLET_PASSWORD,
+                    password_confirm: false,
+                    password_new    : undefined,
+                    password_valid  : undefined
+                });
                 break;
-            case STATUS.NEW_WALLET_CONFIRM_MNEMONIC:
+            case STATUS.NEW_WALLET_PASSWORD:
                 this.createNewWallet();
                 this.setState({status: STATUS.NEW_WALLET_CREATED});
                 break;
@@ -102,19 +77,26 @@ class NewWalletView extends Component {
 
     createWalletPrevStep() {
         switch (this.state.status) {
-            case STATUS.NEW_WALLET_CONFIRM_MNEMONIC:
+            case STATUS.NEW_WALLET_PASSWORD:
                 this.setState({status: STATUS.NEW_WALLET_MNEMONIC});
                 break;
-            case STATUS.NEW_WALLET_MNEMONIC:
-                this.setState({
-                    status          : STATUS.NEW_WALLET_PASSWORD,
-                    password_confirm: false,
-                    password_new    : undefined,
-                    password_valid  : undefined
-                });
-                break;
-            default:
         }
+    }
+
+    getStepName() {
+        const result_step = [
+            {
+                label: Translation.getPhrase('3f1cc1619')
+            },
+            {
+                label: Translation.getPhrase('da19aabd1')
+            },
+            {
+                label: Translation.getPhrase('3f1cc1619')
+            }
+        ];
+
+        return result_step[this.state.status - 1].label;
     }
 
     render() {
@@ -122,29 +104,28 @@ class NewWalletView extends Component {
             return <Redirect to={{pathname: '/'}}/>;
         }
 
-        const {status}             = this.state;
-        const next_button_disabled = !this.state.password_valid ||
-                                     (this.state.status === STATUS.NEW_WALLET_MNEMONIC && !this.state.mnemonic) ||
-                                     (this.state.status === STATUS.NEW_WALLET_CONFIRM_MNEMONIC && !this.state.mnemonic_is_confirmed);
+        const next_button_disabled = !this.state.mnemonic_is_confirmed ||
+                                     (this.state.status === STATUS.NEW_WALLET_PASSWORD && !this.state.password_valid);
 
         let back_button = <Button variant="outline-primary" onClick={() => {
             this.props.history.replace('/unlock-wallet/');
         }}>
             back
         </Button>;
-        if (this.state.status !== STATUS.NEW_WALLET_PASSWORD &&
+        if (this.state.status !== STATUS.NEW_WALLET_MNEMONIC &&
             this.state.status !== STATUS.NEW_WALLET_CREATED) {
             back_button = <Button
-                variant="outline-default"
+                variant="outline-primary"
                 onClick={() => this.createWalletPrevStep()}>back</Button>;
         }
 
+        const {status} = this.state;
         return (
-            <Container className={'create_wallet_container'}>
+            <Container className="import_mnemonic_container">
                 <>
                     <div className={'panel panel-filled'}>
                         <div className={'panel-heading bordered'}>
-                            {Translation.getPhrase('8e0007e24')}
+                            {Translation.getPhrase('e976f2315')}
                         </div>
                         <div className={'panel-body'}>
                             <Row>
@@ -156,23 +137,24 @@ class NewWalletView extends Component {
                                         })} {this.getStepName()}
                                     </div>
                                     <div>
-                                        {this.state.mnemonic && status === STATUS.NEW_WALLET_MNEMONIC && (
-                                            <MnemonicView
-                                                mnemonic={this.state.mnemonic}/>)}
-                                        {this.state.mnemonic && status === STATUS.NEW_WALLET_CONFIRM_MNEMONIC && (
+                                        {status === STATUS.NEW_WALLET_MNEMONIC && (
                                             <MnemonicConfirmView
-                                                processName={'create'}
-                                                mnemonic={this.state.mnemonic}
-                                                onChange={isConfirmed => this.setState({mnemonic_is_confirmed: isConfirmed})}/>)}
-                                        {this.state.status === STATUS.NEW_WALLET_PASSWORD && (
+                                                processName={'import'}
+                                                mnemonic={new Array(24).fill('')}
+                                                importNew={true}
+                                                onChange={(isConfirmed, mnemonic) => this.setState({
+                                                    mnemonic_is_confirmed: isConfirmed,
+                                                    mnemonic
+                                                })}/>)}
+                                        {this.state.mnemonic_is_confirmed && this.state.status === STATUS.NEW_WALLET_PASSWORD && (
                                             <WalletCreatePasswordView
-                                                processName={'create'}
+                                                processName={'import'}
                                                 notConfirmed={!this.state.password_valid}
                                                 onPassword={this.onPassword.bind(this)}
                                                 onConfirmPassword={this.onConfirmPassword.bind(this)}/>)}
                                         {this.state.wallet_info && this.state.status === STATUS.NEW_WALLET_CREATED && (
                                             <WalletCreateInfoView
-                                                processName={'create'}
+                                                processName={'import'}
                                                 wallet={this.state.wallet_info}/>
                                         )}
                                     </div>
@@ -183,7 +165,7 @@ class NewWalletView extends Component {
                                         </div>
                                         <Button variant="outline-primary"
                                                 disabled={next_button_disabled}
-                                                onClick={() => this.createWalletNextStep()}>{Translation.getPhrase('e4c4d7e06')}</Button>
+                                                onClick={() => this.createWalletNextStep()}>{Translation.getPhrase('92a871b62')}</Button>
                                     </div>
                                 </Col>
                             </Row>
@@ -201,4 +183,4 @@ export default connect(
         wallet: state.wallet
     }),
     {unlockWallet}
-)(NewWalletView);
+)(ImportWalletView);
