+++ src/js/components/auth/mnemonic-confirm-view.jsx
