+++ src/js/components/auth/mnemonic-view.jsx
