+++ src/js/components/auth/wallet-create-info-view.jsx
