+++ src/js/components/auth/wallet-create-password-view.jsx
